requirements.txt
aiogram==2.21
python-environ==0.4.54
requests==2.28.1